+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`6.18.0...main`][6.18.0...main].
 ### Changed
 
 - Enabled the `PhpCsFixerCustomFixers/multiline_comment_opening_closing_alone` fixer ([#974]), by [@localheinz]
+- Enabled the `PhpCsFixerCustomFixers/multiline_promoted_properties` fixer ([#975]), by [@localheinz]
 
 ## [`6.18.0`][6.18.0]
 
@@ -1484,6 +1485,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#972]: https://github.com/ergebnis/php-cs-fixer-config/pull/972
 [#973]: https://github.com/ergebnis/php-cs-fixer-config/pull/973
 [#974]: https://github.com/ergebnis/php-cs-fixer-config/pull/974
+[#975]: https://github.com/ergebnis/php-cs-fixer-config/pull/975
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler